requirements.txt
azure-ai-documentintelligence
